+++ src/api/events/useGetEvents.jsx
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
-import getEvents from '../api/getEvents';
-import { useAuth } from '../contexts/authContext';
+import getEvents from './getEvents';
+import { useAuth } from '../../contexts/authContext';
 
 const fetchAllEvents = async (calendars, rooms) => {
   const allEvents = [];
